Bug fixes (#915)
Bug fixes:
- Fix RTL scrolling and alignment
- Fix stale row measurements after keyed reordering
- Fixed dropped prop updates
- Use fractional percentage sizing
- Fix center alignment near list boundaries

Performance tweaks:
- Improve scrolling performance by calculating fixed-size visible ranges directly and searching cached variable-size bounds instead of scanning from the beginning.
- Fixed-size jumps also avoid allocating bounds for preceding items.

+++ lib/core/useItemSize.ts
@@ -20,7 +20,7 @@ export function useItemSize<Props extends object>({
         "Container size must be defined if a percentage item size is specified"
       );
 
-      itemSize = (containerSize * parseInt(itemSizeProp)) / 100;
+      itemSize = (containerSize * parseFloat(itemSizeProp)) / 100;
       break;
     }
     default: {
